requirements.txt
flask
python-telegram-bot==20.3
openai
psycopg2-binary
nest_asyncio